pyproject.toml
[project]
name = "deepagents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "honcho-ai>=1.3.0",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.10",
]
